+++ apps/validator/src/main/scala/org/lfdecentralizedtrust/splice/validator/ValidatorApp.scala
@@ -824,8 +824,6 @@ class ValidatorApp(
             loggerFactory,
             domainMigrationInfo,
             participantId,
-            config.ingestFromParticipantBegin,
-            config.ingestUpdateHistoryFromParticipantBegin,
             config.parameters,
           )
           val walletManager = new UserWalletManager(
@@ -845,8 +843,6 @@ class ValidatorApp(
             loggerFactory,
             domainMigrationInfo,
             participantId,
-            config.ingestFromParticipantBegin,
-            config.ingestUpdateHistoryFromParticipantBegin,
             validatorTopupConfig,
             config.walletSweep,
             config.autoAcceptTransfers,
@@ -889,8 +885,6 @@ class ValidatorApp(
         config.domainMigrationDumpPath,
         config.domainMigrationId,
         retryProvider,
-        config.ingestFromParticipantBegin,
-        config.ingestUpdateHistoryFromParticipantBegin,
         config.svValidator,
         config.sequencerRequestAmplificationPatience,
         config.contactPoint,